refactor: rename files and directories
